requirements.txt
# Clothy 依赖
# Python 3.8+

# 图像处理
pillow>=10.0.0          # PIL License - 图像处理

# HTTP 请求
requests>=2.31.0        # Apache 2.0 - 图片下载、API 调用

# 配置解析
pyyaml>=6.0             # MIT - YAML 配置文件

# 日期处理
python-dateutil>=2.8.0  # Apache 2.0 - 日期解析

# 可选：AI 模型（需要配置 DASHSCOPE_API_KEY）
# 使用阿里云 DashScope API，无需额外安装
# 见 https://dashscope.console.aliyun.com/
